🐛 Fix viewing of events
